Simple label value anonymization with salt
For anonymization, hash designated labels using a cryptographic hash
combined with a local salt that the remote server will not be aware of.
Future approaches may require a whitelist instead of a blacklist for
specifying which labels to transform. Because many input label values
may be guessable, a long salt is recommended.