Add upstream authentication options
The telemeter-server can be configured to forward authentication
requests to an upstream server which may accept or reject a client's
request. The server then returns that info to the client.

Allow the server to serve the --listen address over HTTP or HTTPS
depending on whether a --tls-key and --tls-crt parameter is specified.