jsonnet: telemeter-client fixes
Telemeter client makes no api requests and has no role or role bindings,
so we do not need to allocate a service account.

Openshift does not let you run as user 65534, so let's omit this for
now, we can always add it again later.

The repo for Telemeter is quay.io/openshift/origin-telemeter