jsonnet: reference cluster-monitoring-view
This commit changes the Telemeter client configuration to bind against
an existing `cluster-monitoring-view` ClusterRole, rather than creating
a ClusterRole that duplicates this one. The reasoning for this is that
if the definition of the permissions required for the
`cluster-monitoring-view` change, then the Telemeter code base does not
need to change.