cmd/client,pkg: add reload functionality
This commit restructures the main.go file of the Telemeter client to
move all of the configuration logic for the worker into the `forwarder`
package. This gives us the ability to reload the worker with new
configuration whenever we like.

This commit also adds a reload HTTP endpoint and the ability to reload
configuration on SIGHUP.