rules: Gather cluster version metrics from a cluster
openshift/cluster-version-operator#45 and openshift/cluster-version-operator#47
added a set of core metrics to the cluster version operator to report on the
version of the cluster, the health of the cluster operators, and any payload
errors. As these metrics are of key importance to safe rollout of automated
upgrades to a cluster, gather and report them to telemetry.

These metrics have limited privacy impact when gathering since we only track
version strings. A future release may return the payload image used to carry
the update, which might include local registry information.