jsonnet/telemeter/client: fix configmap volume
This commit fixes the definition of the Volume for the serving certs CA
bundle ConfigMap. Instead of specifying a list of key-path pairs to
load explicitly, we let the Volume load all keys from the ConfigMap.
Unfortunately the Volume constructor for ConfigMaps requires a list of
items, so we bypass the constructor and use the mixins to generate the
volume.