pkg/authorize/jwt: fix handling of required labels
This commit fixes the handling of required labels in the JWT package.
Formerly, the JWT authorizer constructor accepted a labels parameter,
but this map was never actually used by the type. This commit fixes the
handling of the parameter so that the labels are added to the JWT. This
means that the labels will now be respected as required labels in the
validator.

This commit also cleans up some general label handling. For example, the
validator was had the responsibility of adding additional labels
to the received metrics, when this really has nothing to do with
validation; it is instead a general purpose transformation. This has
been extracted and put instead into the `Post` handler so that the label
addition occurs after successful validation.