cmd/telemeter-server: enable internal TLS
This commit adds real support for actually running the internal HTTP
server with TLS. Before, internal TLS was hardcoded to false and there
was no way to supply certificates for the TLS server.

This commit does not turn on TLS for the internal server but makes it
possible. It also eliminates the use of cmux, as the internal and
external servers are only ever serving one single protocol on a given
port.