jsonnet: restructure jsonnet dir
This commit restructures the jsonnet/telemeter directory. The intention
here is to promote easier reuse of shared functions between the
different components. Chiefly, this allows the Telemeter server
libsonnet to reuse the `withNamespace` function. The reason for this
is that we need to substitute the namespace into the ServiceMonitor TLS
config `serverName` fields so that the field matches the SAN of the
injected service serving certificate and Prometheus is able to scrape
the target.