jsonnet: fix prometheus SAR
This commit fixes the definition of the SAR for the proxy sitting in
front of the Telemeter Prometheus. The SAR was specifying an undefined
field `resourceName`, which was turning the SAR into `get namespace *`
rather than `get namespace $(NAMESPACE)`, which would fail.

This commit also removes the `namespace` field from the SAR, since
namespaces themselves are not namespaced objects.