jsonnet: namespace non-namespaced objects
This commit ensures that the non-namespaced objects in the generated
template manifests will not conflict when they are created in the
same cluster. Currently, if we create a ClusterRoleBinding templated for
production and later another for staging, the ClusterRoleBinding for
staging will stomp on the binding against the production ServiceAccount.
This ensures that the names of such objects will include the
namespace name.