Makefile: create generate test
The Telemeter project includes a large number of automatically generated
files. In order to ensure that these files are always up to date, this
commit introduces a test to check that there is no diff after making all
the generated targets.

This commit will be followed up by a pull request to CI configuration
for this repository.