jsonnet: bump server ratelimit
This sets the rate-limit to 100 times the intended frequency.
If a cluster is making requests more than every two seconds,
the requests will fail. This will also allow clusters reporting
on behalf of others, e.g. CI, to make requests more often.
The long-term solution is to introduce per-ID limits, or
a whitelist of IDs that are not rate-limited.