jsonnet: take some values from spec
The Telemeter client should take some non-sensitive values directly from
the pod spec rather than from the secret. This way, if the values in the
pod spec are modified by the cluster-monitoring-operator, the pod will
be rolled. Otherwise, the operator would have to check for a changed
secret and then scale the pod down and up.

Note that introducing a config-reloader would not help in this case as
the values are taken from command-line arguments that are never re-read,
rather than read from files.