jsonnet/telemeter/server: whitelist out of secret
This commit removes the whitelist from the Telemeter server secret and
places it instead in a series of command line `--whitelist` flags. This
is helpful for several reasons:
1. as the whitelist is no longer in a secret it no longer has to be
handled manually, preventing potential human error when the whitelist
changes; and
2. when the whitelist changes, the DaemonSet spec will also change so
the new configuration will automatically be rolled out without needing
to manually kick pods.