test: create benchmarking script
This commit creates the benchmarking test script. This script creates
the benchmarking resources on a cluster, runs the benchmarking binary,
and collects test results.

The results of this benchmarking test can be analyzed using the
newly added test/plot.py script.

Finally, test/timeseries.txt contains example timeseries used by the
benchmark binary to create payloads to send to the Telemeter server
against which the benchmark tests are run.