jsonnet/telemeter/client: move metrics from secret
This commit moves the metric match rules for the client out of the
secret and directly into the container args. This makes it easier to
identify the metrics for anyone inspecting the manifests. This is
desirable in order to:
1. increase trust and transparency
2. decrease the chance for mistakes
3. reduce complexity
Finally, the metrics are not sensitive and did not need to be in the
secret.