Makefile: only benchmark one value
This commit ensures that the benchmark test will begin and end with the
goal value. Formerly, the benchmark test would begin at 1000 and end at
the goal, allowing the runner to find the highest successful value;
however, this takes a very long time. Now, the test will start at 5000
and exit 0 or 1 depending on if the test succeeds.