jsonnet: change benchmark crb name
This commit changes the name of the benchmarking Prometheus Operator
cluster role binding. Today, this cluster role binding has the same name
as the cluster role binding for Prometheus Operator created by the
monitoring stack but a different subject.

The result is that on a regular OpenShift cluster, the
benchmarking CRB will be overwritten by the CMO. When the CRB is
overwritten, the benchmarking Prometheus-Operator no longer has the
correct privileges and crashloops. This causes flakiness in the
benchmarking tests: when the timing is right, the benchmarking
Prometheus Operator is able to create a Prometheus, and when the timing
is wrong, it is not able to create a Prometheus, causing the tests to
fail.