cmd,pkg: eliminate gossip and prometheus fed
This commit eliminates the Prometheus federation portion of Telemeter
server. This effectively means that Telemeter server is now only an
authentication and validation proxy that forwards metrics via remote
write.

It also removes the gossip cluster feature of the project as this is no
longer. Previously it served to ensure that metrics only ended up on one
cluster node so that there would be no duplicate samples when federated
by Prometheus.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>