test/integration: fix already used port race
Occasionally, our integration tests are flaking because a process
listening on a certain port in the first integration test has not exited
before another process tries to listen on that port in the second
integration test. This commit gives the processes in the second
integration test a separate set of ports to ensure that the integration
tests do not conflict with one-another.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>