jsonnet/telemeter: Record which cloudpak a cluster appears to be
All known cloudpaks install at least one of the 'ibm-common-service'
operator or the 'ibm-licensing-operator'. To better track the known
cloudpaks report a metric that encodes the cloudpak_type of a cluster.
For now the only value reported is `unknown` to leave room to detect
type in the future.

A subsequent change will add this to the summary recording rule.