jsonnet/telemeter: Add a rule for normalizing ebs_account data
Combine the class and email_domain fields on subscription labels into
a smaller subset query. Add internal as it is used today for legacy
behavior, but leverage class to better discriminate between different
types of accounts (Internal, Customer, Partner, or Evaluation). The
resulting account_type should be used for classification of the
customer type.