Makefile: fix image location in the CI benchmark job
The IMAGE_FORMAT environment variable has been deprecated and removed
which means that the test-benchmark target has been broken when launched
in a CI environment. We somehow fixed it in [0] but it isn't optimal
since the CI jobs should use the image built from the submitted code
change.

The OpenShift CI tooling allows to inject environment variables
referring to images but it has to be done explicitly [1]. There's a
pending change [2] in github.com/openshift/release to inject the
telemeter image location into the CI_TELEMETER_IMAGE variable.

[0]
https://github.com/openshift/telemeter/commit/85d263366ab63104a988e789cecebbb7727fcf32
[1]
https://docs.ci.openshift.org/docs/architecture/ci-operator/#referring-to-images-in-tests
[2] https://github.com/openshift/release/pull/23052

Signed-off-by: Simon Pasquier <spasquie@redhat.com>