Fix issue in cluster:capacity_effective_cpu_cores
This also addresses a bug where 0 was the effective value due to
label mismatches in `cluster:cpu_capacity_cores:_id`.

When we added a recording rule named `cluster:cpu_capacity_cores:_id`,
the recorded time series includes a tenant_id label and so it's not
equivalent to the expression it replaced.

Adding `tenant_id` to the `sum by` expression makes it properly match
the `cluster:cpu_cpacity_cores:_id` series.